QHDEV-734 - Move modal, cta_links and select_boxes into their own component directories
These three were the only components whose JS lived under
`_global/js/` rather than in a component directory of their own, which
made `_global` look like it owned them. Each now sits at
`src/components/<name>/js/global.js` like every other component.

Pure moves: the module bodies are unchanged, only the import paths in
`component-loader.js`, `legacyGlobal.js` and the stories are rewritten.

Their SCSS still lives under `src/components/_global/css/`, so the
triad is not yet complete for these three. Moving the styles changes
the `dist/main.css` rule order, so it is left to a follow-up that can
be diffed against the built output on its own.
