+++ src/components/modal/js/global.js
