+++ src/stories/GlobalNavigation/MegaMainNavigation/mega_main_navigation.stories.js
@@ -3,7 +3,7 @@ import { initMegaMenu } from "../../../components/mega_main_navigation/js/global
 import { within, userEvent, expect } from "storybook/test";
 import { storyParams } from "../../../../.storybook/globals";
 import { initComponents } from "../../../../.storybook/decorators";
-import initCtaLinks from "../../../components/_global/js/cta_links/global";
+import initCtaLinks from "../../../components/cta_links/js/global";
 import { renderHeader, headerArgs } from "../../Header/Header.js";
 
 function render(args) {
