+++ src/stories/LinkColumns/LinkColumns.stories.js
@@ -2,7 +2,7 @@ import { LinkColumns } from "./LinkColumns";
 import { themes, dummyLink, storyParams } from "../../../.storybook/globals";
 import { themeWrapper } from "../../../.storybook/helper-functions";
 import { initComponents } from "../../../.storybook/decorators";
-import initCtaLinks from "../../components/_global/js/cta_links/global";
+import initCtaLinks from "../../components/cta_links/js/global";
 
 const linkColumnsArgs = {
   ariaLabel: "related links",
